+++ pvm/ts/src/commands/mintCommand.test.ts
@@ -1,14 +1,14 @@
 import { ethers, JsonRpcProvider, Contract } from "ethers";
 import { MintCommand } from "./mintCommand";
 import { getMempoolInstance } from "../core/mempool";
-import { getTransactionInstance } from "../state/transactionManagement";
+import { getTransactionInstance } from "../state/index";
 import { Transaction } from "../models/transaction";
 import { NativeToken } from "../models/nativeToken";
 
 // Mock external dependencies
 jest.mock("ethers");
 jest.mock("../core/mempool");
-jest.mock("../state/transactionManagement");
+jest.mock("../state/index");
 jest.mock("../models/transaction");
 jest.mock("../core/provider");
